requirements.txt
asgiref==3.3.1
Django==3.1.6
django-filter==2.4.0
django-js-asset==1.2.2
djangorestframework==3.12.2
Pillow==8.1.0
pkg-resources==0.0.0
python-slugify==4.0.1
pytz==2021.1
sqlparse==0.4.1


